Tests/LibWeb/Text/expected/css/pseudo-element-hover-invalidation.txt

initial color: rgb(255, 0, 0)
hovered color: rgb(0, 128, 0)
unhovered color: rgb(255, 0, 0)
